requirements.txt
dj-database-url==0.5.0
Django==3.0.7
django-filter==2.3.0
django-quill-editor==0.1.16
django-quilljs==0.0.2
django-tempus-dominus==5.1.2.13
django-widget-tweaks==1.4.8
gunicorn==20.0.4
psycopg2-binary==2.7.7
whitenoise==5.1.0
